Separated redux actions from reducers, changed Fetch API to Axios
