+++ EventAsker.SPA/src/actions/types.js
