+++ EventAsker.SPA/src/index.js
@@ -10,10 +10,10 @@ import thunk from 'redux-thunk';
 import { createStore, applyMiddleware} from 'redux';
 import { composeWithDevTools } from 'redux-devtools-extension';
 import rootReducer from './rootReducer';
-import { setCurrentUser } from './helpers/authManager';
+import { setCurrentUser } from './actions/userAction';
 import jwt from 'jsonwebtoken';
 
-const store = createStore( rootReducer,
+const store = createStore(rootReducer,
     composeWithDevTools(
         applyMiddleware(thunk)
     )
